pyproject.toml
[project]
name = "sennen"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
  "openai",
  "pyyaml",
]

[dependency-groups]
dev = [
  "pytest",
  "requests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-m 'not network and not claude'"
markers = [
  "network: tests that hit external APIs (deselect with -m 'not network')",
  "claude: tests that invoke the claude CLI (deselect with -m 'not claude')",
  "llm: tests that call an LLM API (deselect with -m 'not llm')",
  "slow: tests that take more than a few seconds",
]
